requirements.txt
requests~=2.27.1
tldextract~=3.1.2
psycopg2-binary~=2.9.3
beautifulsoup4~=4.10.0
lxml~=4.7.1
pysocks~=1.7.1